refactor(cli): move publish commands to sdk
