+++ packages/cli/tests/e2e/e2e.nonQv.test.ts
@@ -11,11 +11,12 @@ import {
   setVerifyingKeys,
   EMode,
   proveOnChain,
+  publish,
 } from "maci-sdk";
 
 import type { Signer } from "ethers";
 
-import { deploy, deployPoll, deployVkRegistryContract, genProofsCommand, publish, timeTravel } from "../../ts/commands";
+import { deploy, deployPoll, deployVkRegistryContract, genProofsCommand, timeTravel } from "../../ts/commands";
 import { DEFAULT_SG_DATA, DeployedContracts, GenProofsArgs } from "../../ts/utils";
 import {
   deployPollArgs,
