+++ packages/cli/ts/commands/index.ts
@@ -1,7 +1,6 @@
 export { deploy } from "./deploy";
 export { deployPoll } from "./deployPoll";
 export { deployVkRegistryContract } from "./deployVkRegistry";
-export { publish, publishBatch } from "./publish";
 export { showContracts } from "./showContracts";
 export { timeTravel } from "./timeTravel";
 export { genProofsCommand } from "./genProofs";
