--- packages/cli/ts/sdk/index.ts
@@ -1,5 +1,0 @@
-import { publish, publishBatch } from "../commands/publish";
-
-export { publish, publishBatch };
-
-export type { PublishArgs, IPublishBatchArgs, IPublishBatchData, IPublishMessage } from "../utils";
