+++ packages/cli/ts/utils/interfaces.ts
@@ -1,6 +1,5 @@
 import type { Signer } from "ethers";
 import type { CircuitInputs } from "maci-core";
-import type { IMessageContractParams } from "maci-domainobjs";
 import type { SnarkProof } from "maci-sdk";
 import type { Groth16Proof, PublicSignals } from "snarkjs";
 
@@ -310,131 +309,6 @@ export interface GenProofsArgs {
    * Backup files for ipfs messages (name format: ipfsHash.json)
    */
   ipfsMessageBackupFiles?: string[];
-}
-
-/**
- * Interface for the arguments to the publish command
- */
-export interface PublishArgs extends IPublishMessage {
-  /**
-   * The poll public key
-   */
-  pubkey: string;
-
-  /**
-   * The poll private key
-   */
-  privateKey: string;
-
-  /**
-   * The address of the MACI contract
-   */
-  maciAddress: string;
-
-  /**
-   * The id of the poll
-   */
-  pollId: bigint;
-
-  /**
-   * A signer object
-   */
-  signer: Signer;
-
-  /**
-   * Whether to log the output
-   */
-  quiet?: boolean;
-}
-
-/**
- * Interface for the arguments to the batch publish command
- */
-export interface IPublishBatchArgs {
-  /**
-   * User messages
-   */
-  messages: IPublishMessage[];
-
-  /**
-   * The id of the poll
-   */
-  pollId: bigint;
-
-  /**
-   * The address of the MACI contract
-   */
-  maciAddress: string;
-
-  /**
-   * The public key of the user
-   */
-  publicKey: string;
-
-  /**
-   * The private key of the user
-   */
-  privateKey: string;
-
-  /**
-   * A signer object
-   */
-  signer: Signer;
-
-  /**
-   * Whether to log the output
-   */
-  quiet?: boolean;
-}
-
-/**
- * Interface that represents user publish message
- */
-export interface IPublishMessage {
-  /**
-   * The index of the state leaf
-   */
-  stateIndex: bigint;
-
-  /**
-   * The index of the vote option
-   */
-  voteOptionIndex: bigint;
-
-  /**
-   * The nonce of the message
-   */
-  nonce: bigint;
-
-  /**
-   * The new vote weight
-   */
-  newVoteWeight: bigint;
-
-  /**
-   * The salt of the message
-   */
-  salt?: bigint;
-}
-
-/**
- * Interface that represents publish batch return data
- */
-export interface IPublishBatchData {
-  /**
-   * Publish transaction hash
-   */
-  hash?: string;
-
-  /**
-   * Encrypted publish messages
-   */
-  encryptedMessages: IMessageContractParams[];
-
-  /**
-   * Encryption private keys
-   */
-  privateKeys: string[];
 }
 
 /**
