+++ packages/sdk/ts/vote/index.ts
@@ -1,4 +1,15 @@
-export type { IVote } from "./types";
+export type {
+  IVote,
+  IGenerateVoteArgs,
+  IPublishArgs,
+  IPublishBatchArgs,
+  IPublishBatchData,
+  IPublishData,
+  IPublishMessage,
+  ISubmitVoteArgs,
+  ISubmitVoteBatchArgs,
+} from "./types";
 export { generateVote } from "./generate";
+export { publish, publishBatch } from "./publish";
 export { submitVote, submitVoteBatch } from "./submit";
 export { getCoordinatorPubKey } from "./utils";
